Refactor out the components folder. All components live in .
